Base for pure C++ NCCL interface
